prototype/examples/accessing_api/requirements.txt
flask
requests
certifi
pyyaml
websockets
